refactor(storage): split videos.ts into focused modules
Break the 1195-line storageService/videos.ts into sibling modules:
videoQueries (getVideos/getVideoById/getVideoBySourceUrl + visitor
filter), videoVisibility (isVideoPublic/classifyMediaVisibility/cloud
thumbnail rules), videoMutations (save/update + insert flag), videoDeletion
(deleteVideo + file/thumbnail/avatar/subtitle cleanup +
isThumbnailReferencedByOtherVideo), and legacyFilenames (formatLegacyFilenames).

videos.ts is now a ~30-line barrel re-exporting the same public surface,
so the storageService index and all consumers are unaffected. Behavior
is unchanged, including the nosemgrep annotations and GHSA visibility
comments.
