+++ backend/src/services/storageService/videos.ts
@@ -1,1195 +1,29 @@
-import { and, desc, eq, isNull, or, sql, type SQL } from "drizzle-orm";
-import crypto from "crypto";
-import path from "path";
-import {
-    AVATARS_DIR,
-    IMAGES_DIR,
-    SUBTITLES_DIR,
-    UPLOADS_DIR,
-    VIDEOS_DIR,
-} from "../../config/paths";
-import { db } from "../../db";
-import { videos } from "../../db/schema";
-import { DatabaseError } from "../../errors/DownloadErrors";
-import { resolveManagedWebPath } from "../filenameTemplate/pathHelpers";
-import { removeMediaServerArtifactsForVideo } from "../mediaServerExport";
-import { formatVideoFilename } from "../../utils/helpers";
-import { logger } from "../../utils/logger";
-import { getCollections } from "./collections";
-import {
-  buildStoragePath,
-  findImageFile,
-  findVideoFile,
-  findVideoFilesByFilename,
-  pathExists,
-  removeEmptyDirectoryChain,
-  removeFileIfExists,
-  renamePath,
-} from "./fileHelpers";
-import { markDownloadHistoryDeletedByVideoId } from "./downloadHistory";
-import { markVideoDownloadDeleted } from "./videoDownloadTracking";
-import {
-  deleteSmallThumbnailMirrorSync,
-  moveSmallThumbnailMirrorSync,
-  resolveManagedThumbnailWebPathFromAbsolutePath,
-} from "../thumbnailMirrorService";
-
-export type VideoCallerRole = "admin" | "visitor";
-
-// Visibility: 0 = hidden, 1 = public (see db/schema.ts). Visitors are only
-// allowed to see public videos; admins see everything. Mirrors the existing
-// RSS feed filter (rssService.ts). Used to fix GHSA-hcm6-w6x8-6jhr.
-// A NULL visibility is treated as public to match the `(visibility ?? 1) === 1`
-// contract in isVideoPublic and the media classifier; a plain `= 1` would hide
-// imported/legacy rows that wrote NULL from visitors only.
-// Built lazily so the schema reference is not evaluated at module load time
-// (which would break tests that partially mock the db schema).
-const publicOnlyVisitorFilter = () =>
-  or(eq(videos.visibility, 1), isNull(videos.visibility));
-
-export function getVideos(
-  role?: VideoCallerRole
-): import("./types").Video[] {
-  try {
-    const baseQuery = db
-      .select()
-      .from(videos)
-      .orderBy(desc(videos.createdAt));
-    const allVideos =
-      role === "visitor"
-        ? baseQuery.where(publicOnlyVisitorFilter()).all()
-        : baseQuery.all();
-    return allVideos.map((v) => ({
-      ...v,
-      tags: v.tags ? JSON.parse(v.tags) : [],
-      subtitles: v.subtitles ? JSON.parse(v.subtitles) : undefined,
-    })) as import("./types").Video[];
-  } catch (error) {
-    logger.error(
-      "Error getting videos",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    // Return empty array for backward compatibility with frontend
-    return [];
-  }
-}
-
-export function getVideoBySourceUrl(
-  sourceUrl: string
-): import("./types").Video | undefined {
-  try {
-    const result = db
-      .select()
-      .from(videos)
-      .where(eq(videos.sourceUrl, sourceUrl))
-      .get();
-
-    if (result) {
-      return {
-        ...result,
-        tags: result.tags ? JSON.parse(result.tags) : [],
-        subtitles: result.subtitles ? JSON.parse(result.subtitles) : undefined,
-      } as import("./types").Video;
-    }
-    return undefined;
-  } catch (error) {
-    logger.error(
-      "Error getting video by sourceUrl",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to get video by source URL: ${sourceUrl}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "getVideoBySourceUrl"
-    );
-  }
-}
-
-export function getVideoById(
-  id: string,
-  role?: VideoCallerRole
-): import("./types").Video | undefined {
-  try {
-    const baseQuery = db.select().from(videos);
-    const video =
-      role === "visitor"
-        ? baseQuery
-            .where(
-              and(eq(videos.id, id), publicOnlyVisitorFilter())
-            )
-            .get()
-        : baseQuery.where(eq(videos.id, id)).get();
-    if (video) {
-      return {
-        ...video,
-        tags: video.tags ? JSON.parse(video.tags) : [],
-        subtitles: video.subtitles ? JSON.parse(video.subtitles) : undefined,
-      } as import("./types").Video;
-    }
-    return undefined;
-  } catch (error) {
-    logger.error(
-      "Error getting video by id",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to get video by id: ${id}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "getVideoById"
-    );
-  }
-}
-
-/**
- * A video is treated as publicly visible unless visibility is explicitly 0.
- * Mirrors the frontend contract `(visibility ?? 1) === 1`.
- */
-export function isVideoPublic(video: {
-  visibility?: number | null;
-}): boolean {
-  return (video.visibility ?? 1) === 1;
-}
-
-export type MediaVisibility = "public" | "hidden" | "unknown";
-
-const escapeLikePattern = (value: string): string =>
-  value.replace(/[\\%_]/g, "\\$&");
-
-const CLOUD_THUMBNAIL_CACHE_KEY_PATTERN = /^[a-f0-9]{64}\.jpg$/;
-
-const normalizeCloudThumbnailCacheKey = (value: string): string | null => {
-  const key = path.posix.basename(value).toLowerCase();
-  return CLOUD_THUMBNAIL_CACHE_KEY_PATTERN.test(key) ? key : null;
-};
-
-const getCloudThumbnailCacheKey = (cloudPath: string): string =>
-  `${crypto.createHash("sha256").update(cloudPath).digest("hex")}.jpg`;
-
-const getCachedCloudThumbnailVisibilityRows = (
-  cacheKeys: Set<string>
-): Array<{ visibility: number | null }> => {
-  if (cacheKeys.size === 0) {
-    return [];
-  }
-
-  const rows = db
-    .select({
-      thumbnailPath: videos.thumbnailPath,
-      thumbnailUrl: videos.thumbnailUrl,
-      visibility: videos.visibility,
-    })
-    .from(videos)
-    .where(
-      or(
-        sql`${videos.thumbnailPath} LIKE ${"cloud:%"}`,
-        sql`${videos.thumbnailUrl} LIKE ${"cloud:%"}`
-      )
-    )
-    .all();
-
-  return rows
-    .filter((row) => {
-      const candidates = [row.thumbnailPath, row.thumbnailUrl].filter(
-        (value): value is string => typeof value === "string" && value.length > 0
-      );
-      return candidates.some((candidate) =>
-        cacheKeys.has(getCloudThumbnailCacheKey(candidate))
-      );
-    })
-    .map((row) => ({ visibility: row.visibility }));
-};
-
-/**
- * Classify a piece of static media by the visibility of the video(s) that
- * reference it. Used as defense-in-depth on the static/cloud media routes
- * (GHSA-hcm6-w6x8-6jhr): even after the login wall, media that belongs only to
- * hidden videos must not be served to non-admin callers, while public media
- * stays reachable (e.g. for RSS clients).
- *
- * - `exactPaths` are matched against `videoPath`/`thumbnailPath`/`thumbnailUrl`.
- * - `subtitlePaths` are matched against entries inside the `subtitles` JSON.
- * - `cloudThumbnailCacheKeys` are matched by recomputing the stable cache key
- *   for cloud thumbnail paths, which lets the static cache route enforce the
- *   same visibility rule even though its URL only contains a hash filename.
- *
- * Returns "public" if any referencing video is visible, "hidden" if every
- * referencing video is hidden, and "unknown" if no video references the media
- * (orphan files, frontend assets, avatars — left untouched).
- */
-export function classifyMediaVisibility(opts: {
-  exactPaths?: string[];
-  subtitlePaths?: string[];
-  cloudThumbnailCacheKeys?: string[];
-}): MediaVisibility {
-  const conditions: SQL[] = [];
-  const cloudThumbnailCacheKeys = new Set(
-    (opts.cloudThumbnailCacheKeys ?? [])
-      .map(normalizeCloudThumbnailCacheKey)
-      .filter((key): key is string => key != null)
-  );
-
-  for (const candidate of opts.exactPaths ?? []) {
-    if (!candidate) continue;
-    conditions.push(eq(videos.videoPath, candidate));
-    conditions.push(eq(videos.thumbnailPath, candidate));
-    conditions.push(eq(videos.thumbnailUrl, candidate));
-  }
-
-  for (const candidate of opts.subtitlePaths ?? []) {
-    if (!candidate) continue;
-    const pattern = `%"${escapeLikePattern(candidate)}"%`;
-    conditions.push(sql`${videos.subtitles} LIKE ${pattern} ESCAPE '\\'`);
-  }
-
-  if (conditions.length === 0 && cloudThumbnailCacheKeys.size === 0) {
-    return "unknown";
-  }
-
-  try {
-    const rows: Array<{ visibility: number | null }> =
-      conditions.length > 0
-        ? db
-            .select({ visibility: videos.visibility })
-            .from(videos)
-            .where(or(...conditions))
-            .all()
-        : [];
-
-    rows.push(...getCachedCloudThumbnailVisibilityRows(cloudThumbnailCacheKeys));
-
-    if (rows.length === 0) {
-      return "unknown";
-    }
-
-    // Any visible reference makes the file public (a thumbnail could be shared);
-    // only block when every match is hidden.
-    return rows.some((row) => isVideoPublic(row)) ? "public" : "hidden";
-  } catch (error) {
-    logger.error(
-      "Error classifying media visibility",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    // Fail closed: treat as hidden so a transient DB error can't leak media.
-    return "hidden";
-  }
-}
-
-/**
- * Whether a visitor may access a cloud-stored file. Shared by the cloud
- * redirect routes and the signed-url controller so the "is this `cloud:` ref
- * hidden?" rule lives in exactly one place. A file is visitor-accessible unless
- * every video referencing it is hidden; orphan files (no match) stay reachable
- * so cached thumbnails keep resolving. Part of GHSA-hcm6-w6x8-6jhr.
- */
-export function isCloudFileVisibleToVisitor(filename: string): boolean {
-  if (!filename) return true;
-  return (
-    classifyMediaVisibility({ exactPaths: [`cloud:${filename}`] }) !== "hidden"
-  );
-}
-
-/**
- * Check if a video part already exists by sourceUrl
- * Returns the existing video if found, undefined otherwise
- */
-export function getVideoPartBySourceUrl(
-  sourceUrl: string
-): import("./types").Video | undefined {
-  return getVideoBySourceUrl(sourceUrl);
-}
-
-/**
- * Format legacy filenames to the new standard format: Title-Author-YYYY
- */
-export function formatLegacyFilenames(): {
-  processed: number;
-  renamed: number;
-  errors: number;
-  details: string[];
-} {
-  const results = {
-    processed: 0,
-    renamed: 0,
-    errors: 0,
-    details: [] as string[],
-  };
-
-  try {
-    const allVideos = getVideos();
-    logger.info(
-      `Starting legacy filename formatting for ${allVideos.length} videos...`
-    );
-
-    for (const video of allVideos) {
-      results.processed++;
-
-      try {
-        // Generate new filename
-        const newBaseFilename = formatVideoFilename(
-          video.title,
-          video.author || "Unknown",
-          video.date
-        );
-
-        // preserve subdirectory if it exists (e.g. for collections)
-        // We rely on videoPath because videoFilename is usually just the basename
-        let subdirectory = "";
-        if (video.videoPath) {
-          // videoPath is like "/videos/SubDir/file.mp4" or "/videos/file.mp4"
-          const relPath = video.videoPath.replace(/^\/videos\//, "");
-          const dir = path.dirname(relPath);
-          if (dir && dir !== ".") {
-            subdirectory = dir;
-          }
-        }
-
-        // New filename (basename only)
-        const newVideoFilename = `${newBaseFilename}.mp4`;
-        const newThumbnailFilename = `${newBaseFilename}.jpg`;
-
-        // Calculate full paths for checks
-        // For the check we need to know if the resulting full path is different
-        // But the check "video.videoFilename === newVideoFilename" only checks basename.
-        // If basename matches, we might still want to rename if we were normalizing something else,
-        // but usually if format matches, we skip.
-        if (video.videoFilename === newVideoFilename) {
-          continue;
-        }
-
-        logger.info(
-          `Renaming video ${video.id}: ${video.videoFilename} -> ${newVideoFilename} (Subdir: ${subdirectory})`
-        );
-
-        // Paths
-        // Old path must be constructed using the subdirectory derived from videoPath
-        const oldVideoPath = buildStoragePath(
-          VIDEOS_DIR,
-          subdirectory,
-          video.videoFilename || ""
-        );
-        const newVideoPath = buildStoragePath(
-          VIDEOS_DIR,
-          subdirectory,
-          newVideoFilename
-        );
-
-        // Handle thumbnail subdirectory
-        let thumbSubdir = "";
-        let thumbnailBaseDir = IMAGES_DIR;
-        let thumbnailPathPrefix = "/images";
-        if (video.thumbnailPath) {
-          const relPath = video.thumbnailPath
-            .replace(/^\/images\//, "")
-            .replace(/^\/videos\//, "");
-          const dir = path.dirname(relPath);
-          if (dir && dir !== ".") {
-            thumbSubdir = dir;
-          }
-
-          if (video.thumbnailPath.startsWith("/videos/")) {
-            thumbnailBaseDir = VIDEOS_DIR;
-            thumbnailPathPrefix = "/videos";
-          }
-        }
-
-        const oldThumbnailPath = video.thumbnailFilename
-          ? buildStoragePath(
-              thumbnailBaseDir,
-              thumbSubdir,
-              video.thumbnailFilename,
-            )
-          : null;
-        const newThumbnailPath = buildStoragePath(
-          thumbnailBaseDir,
-          thumbSubdir,
-          newThumbnailFilename
-        );
-
-        // Rename video file
-        if (pathExists(oldVideoPath)) {
-          if (pathExists(newVideoPath) && oldVideoPath !== newVideoPath) {
-            // Destination exists, append timestamp to avoid collision
-            const uniqueSuffix = `_${Date.now()}`;
-            const uniqueBase = `${newBaseFilename}${uniqueSuffix}`;
-
-            const uniqueVideoBase = `${uniqueBase}.mp4`;
-            const uniqueThumbBase = `${uniqueBase}.jpg`;
-
-            // Full paths for rename
-            const uniqueVideoPath = buildStoragePath(
-              VIDEOS_DIR,
-              subdirectory,
-              uniqueVideoBase
-            );
-            const uniqueThumbPath = buildStoragePath(
-              thumbnailBaseDir,
-              thumbSubdir,
-              uniqueThumbBase
-            ); // Use thumbSubdir
-
-            logger.info(
-              `Destination exists, using unique suffix: ${uniqueVideoBase}`
-            );
-
-            renamePath(oldVideoPath, uniqueVideoPath);
-
-            if (oldThumbnailPath && pathExists(oldThumbnailPath)) {
-              renamePath(oldThumbnailPath, uniqueThumbPath);
-              moveSmallThumbnailMirrorSync(
-                video.thumbnailPath,
-                `${thumbnailPathPrefix}/${
-                  thumbSubdir ? thumbSubdir + "/" : ""
-                }${uniqueThumbBase}`,
-              );
-            }
-
-            // Handle subtitles (Keep in their original folder, assuming root or derived from path if available)
-            if (video.subtitles && video.subtitles.length > 0) {
-              const newSubtitles = [];
-              for (const subtitle of video.subtitles) {
-                // Subtitles usually in SUBTITLES_DIR root, checking...
-                const oldSubPath = buildStoragePath(
-                  SUBTITLES_DIR,
-                  subtitle.filename,
-                );
-
-                // If we ever supported subdirs for subtitles, we'd need to parse subtitle.path here too
-                // For now assuming existing structure matches simple join
-
-                if (pathExists(oldSubPath)) {
-                  const newSubFilename = `${uniqueBase}.${subtitle.language}.vtt`;
-                  const newSubPath = buildStoragePath(
-                    SUBTITLES_DIR,
-                    newSubFilename,
-                  );
-                  renamePath(oldSubPath, newSubPath);
-                  newSubtitles.push({
-                    ...subtitle,
-                    filename: newSubFilename,
-                    path: `/subtitles/${newSubFilename}`,
-                  });
-                } else {
-                  newSubtitles.push(subtitle);
-                }
-              }
-              // Update video record with unique names
-              // videoFilename should be BASENAME only
-              // videoPath should be FULL WEB PATH including subdir
-              db.update(videos)
-                .set({
-                  videoFilename: uniqueVideoBase,
-                  thumbnailFilename: video.thumbnailFilename
-                    ? uniqueThumbBase
-                    : undefined,
-                  videoPath: `/videos/${
-                    subdirectory ? subdirectory + "/" : ""
-                  }${uniqueVideoBase}`,
-                  thumbnailPath: video.thumbnailFilename
-                    ? `${thumbnailPathPrefix}/${
-                        thumbSubdir ? thumbSubdir + "/" : ""
-                      }${uniqueThumbBase}`
-                    : null,
-                  subtitles: JSON.stringify(newSubtitles),
-                })
-                .where(eq(videos.id, video.id))
-                .run();
-            } else {
-              // Update video record with unique names
-              db.update(videos)
-                .set({
-                  videoFilename: uniqueVideoBase,
-                  thumbnailFilename: video.thumbnailFilename
-                    ? uniqueThumbBase
-                    : undefined,
-                  videoPath: `/videos/${
-                    subdirectory ? subdirectory + "/" : ""
-                  }${uniqueVideoBase}`,
-                  thumbnailPath: video.thumbnailFilename
-                    ? `${thumbnailPathPrefix}/${
-                        thumbSubdir ? thumbSubdir + "/" : ""
-                      }${uniqueThumbBase}`
-                    : null,
-                })
-                .where(eq(videos.id, video.id))
-                .run();
-            }
-
-            results.renamed++;
-            results.details.push(`Renamed (unique): ${video.title}`);
-          } else {
-            // Rename normally
-            renamePath(oldVideoPath, newVideoPath);
-
-            if (oldThumbnailPath && pathExists(oldThumbnailPath)) {
-              // Check if new thumbnail path exists (it shouldn't if specific to this video, but safety check)
-              if (pathExists(newThumbnailPath) && oldThumbnailPath !== newThumbnailPath) {
-                removeFileIfExists(newThumbnailPath);
-              }
-              renamePath(oldThumbnailPath, newThumbnailPath);
-              moveSmallThumbnailMirrorSync(
-                video.thumbnailPath,
-                `${thumbnailPathPrefix}/${
-                  thumbSubdir ? thumbSubdir + "/" : ""
-                }${newThumbnailFilename}`,
-              );
-            }
-
-            // Handle subtitles
-            const updatedSubtitles = [];
-            if (video.subtitles && video.subtitles.length > 0) {
-              for (const subtitle of video.subtitles) {
-                const oldSubPath = buildStoragePath(
-                  SUBTITLES_DIR,
-                  subtitle.filename,
-                );
-                if (pathExists(oldSubPath)) {
-                  // Keep subtitles in their current location (usually root SUBTITLES_DIR)
-                  const newSubFilename = `${newBaseFilename}.${subtitle.language}.vtt`;
-                  const newSubPath = buildStoragePath(
-                    SUBTITLES_DIR,
-                    newSubFilename,
-                  );
-
-                  // Remove dest if exists
-                  removeFileIfExists(newSubPath);
-
-                  renamePath(oldSubPath, newSubPath);
-                  updatedSubtitles.push({
-                    ...subtitle,
-                    filename: newSubFilename,
-                    path: `/subtitles/${newSubFilename}`,
-                  });
-                } else {
-                  updatedSubtitles.push(subtitle);
-                }
-              }
-            }
-
-            // Update DB
-            db.update(videos)
-              .set({
-                videoFilename: newVideoFilename,
-                thumbnailFilename: video.thumbnailFilename
-                  ? newThumbnailFilename
-                  : undefined,
-                videoPath: `/videos/${
-                  subdirectory ? subdirectory + "/" : ""
-                }${newVideoFilename}`,
-                thumbnailPath: video.thumbnailFilename
-                  ? `${thumbnailPathPrefix}/${
-                      thumbSubdir ? thumbSubdir + "/" : ""
-                    }${newThumbnailFilename}`
-                  : null,
-                subtitles:
-                  updatedSubtitles.length > 0
-                    ? JSON.stringify(updatedSubtitles)
-                    : video.subtitles
-                    ? JSON.stringify(video.subtitles)
-                    : undefined,
-              })
-              .where(eq(videos.id, video.id))
-              .run();
-
-            results.renamed++;
-          }
-        } else {
-          results.details.push(`Skipped (file missing): ${video.title}`);
-          // results.errors++; // Not necessarily an error, maybe just missing file
-        }
-      } catch (err: any) {
-        logger.error(
-          `Error renaming video ${video.id}`,
-          err instanceof Error ? err : new Error(String(err))
-        );
-        results.errors++;
-        results.details.push(`Error: ${video.title} - ${err.message}`);
-      }
-    }
-
-    return results;
-  } catch (error: any) {
-    logger.error(
-      "Error in formatLegacyFilenames",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      "Failed to format legacy filenames",
-      error instanceof Error ? error : new Error(String(error)),
-      "formatLegacyFilenames"
-    );
-  }
-}
-
-export interface SaveVideoOptions {
-  // When true, suppress the library_video_added statistics event even on insert
-  // (used by maintenance flows like database migration/import/merge that should
-  // not appear as user-facing library mutations).
-  suppressStatistics?: boolean;
-  // Optional reason bucket for the library_video_added event.
-  statisticsReason?: "manual" | "download" | "upload" | "scan" | "subscription" | "task";
-}
-
-export function saveVideo(
-  videoData: import("./types").Video,
-  options: SaveVideoOptions = {}
-): import("./types").Video {
-  const result = saveVideoWithInsertFlag(videoData);
-  if (result.inserted && !options.suppressStatistics) {
-    emitLibraryVideoAdded(videoData, options.statisticsReason ?? "download");
-  }
-  return videoData;
-}
-
-// Same as saveVideo but reports whether a new row was created and never emits.
-// Statistics callers use this directly when they need to control emission timing.
-export function saveVideoWithInsertFlag(
-  videoData: import("./types").Video
-): { video: import("./types").Video; inserted: boolean } {
-  try {
-    const existing = db
-      .select({ id: videos.id })
-      .from(videos)
-      .where(eq(videos.id, videoData.id))
-      .all();
-    const inserted = existing.length === 0;
-    const videoToSave = {
-      ...videoData,
-      tags: videoData.tags ? JSON.stringify(videoData.tags) : undefined,
-      subtitles: videoData.subtitles
-        ? JSON.stringify(videoData.subtitles)
-        : undefined,
-    };
-    db.insert(videos)
-      .values(videoToSave as any)
-      .onConflictDoUpdate({
-        target: videos.id,
-        set: videoToSave,
-      })
-      .run();
-    return { video: videoData, inserted };
-  } catch (error) {
-    logger.error(
-      "Error saving video",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to save video: ${videoData.id}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "saveVideo"
-    );
-  }
-}
-
-function emitLibraryVideoAdded(
-  videoData: import("./types").Video,
-  reason: string
-): void {
-  try {
-    const { recordEvent } = require("../statistics") as {
-      recordEvent: (input: any) => string | null;
-    };
-    const fileSizeBytes =
-      videoData.fileSize !== undefined && videoData.fileSize !== null
-        ? Number(videoData.fileSize) || 0
-        : 0;
-    recordEvent({
-      eventType: "library_video_added",
-      actorRole: "system",
-      surface: "background",
-      videoId: videoData.id,
-      platform:
-        typeof videoData.source === "string"
-          ? videoData.source.toLowerCase()
-          : null,
-      payload: {
-        reason,
-        videoTitle: videoData.title,
-        fileSizeBytes,
-      },
-    });
-  } catch {
-    // statistics is best-effort
-  }
-}
-
-export function saveVideoIfAbsent(
-  videoData: import("./types").Video
-): boolean {
-  try {
-    const videoToSave = {
-      ...videoData,
-      tags: videoData.tags ? JSON.stringify(videoData.tags) : undefined,
-      subtitles: videoData.subtitles
-        ? JSON.stringify(videoData.subtitles)
-        : undefined,
-    };
-
-    const result = db
-      .insert(videos)
-      .values(videoToSave as any)
-      .onConflictDoNothing({
-        target: videos.id,
-      })
-      .run();
-
-    return result.changes > 0;
-  } catch (error) {
-    logger.error(
-      "Error saving video if absent",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to save video if absent: ${videoData.id}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "saveVideoIfAbsent"
-    );
-  }
-}
-
-export function updateVideo(
-  id: string,
-  updates: Partial<import("./types").Video>
-): import("./types").Video | null {
-  try {
-    const updatesToSave: Record<string, unknown> = {
-      ...updates,
-      // Only include tags/subtitles if they are explicitly in the updates object
-      ...(updates.tags !== undefined
-        ? { tags: updates.tags ? JSON.stringify(updates.tags) : undefined }
-        : {}),
-      ...(updates.subtitles !== undefined
-        ? {
-            subtitles: updates.subtitles
-              ? JSON.stringify(updates.subtitles)
-              : undefined,
-          }
-        : {}),
-    };
-
-    // Keep thumbnailUrl aligned when callers only update thumbnailPath
-    if (
-      Object.prototype.hasOwnProperty.call(updates, "thumbnailPath") &&
-      !Object.prototype.hasOwnProperty.call(updates, "thumbnailUrl")
-    ) {
-      updatesToSave.thumbnailUrl = updates.thumbnailPath ?? null;
-    }
-
-    // If tags is explicitly empty array, we might want to save it as '[]' or null.
-    // JSON.stringify([]) is '[]', which is fine.
-
-    const result = db
-      .update(videos)
-      .set(updatesToSave as any)
-      .where(eq(videos.id, id))
-      .returning()
-      .get();
-
-    if (result) {
-      return {
-        ...result,
-        tags: result.tags ? JSON.parse(result.tags) : [],
-        subtitles: result.subtitles ? JSON.parse(result.subtitles) : undefined,
-      } as import("./types").Video;
-    }
-    return null;
-  } catch (error) {
-    logger.error(
-      "Error updating video",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to update video: ${id}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "updateVideo"
-    );
-  }
-}
-
-export function deleteVideo(
-  id: string,
-  reason: "manual" | "retention" | "maintenance" = "manual"
-): boolean {
-  try {
-    const videoToDelete = getVideoById(id);
-    if (!videoToDelete) return false;
-
-    const allCollections = getCollections();
-
-    // Remove video file
-    deleteVideoFile(videoToDelete, allCollections);
-
-    // Remove thumbnail file
-    deleteThumbnailFile(videoToDelete, allCollections);
-
-    // Remove author avatar file only if this is the last video from this author
-    deleteAuthorAvatarIfNeeded(videoToDelete, id);
-
-    // Remove subtitle files
-    deleteSubtitleFiles(videoToDelete, allCollections);
-
-    const deletedAt = Date.now();
-
-    // Mark download tracking and history as deleted
-    markVideoDownloadDeleted(id);
-    markDownloadHistoryDeletedByVideoId(id, deletedAt);
-
-    // Delete from DB
-    db.delete(videos).where(eq(videos.id, id)).run();
-    removeMediaServerArtifactsForVideo(videoToDelete, {
-      libraryVideos: getVideos(),
-    });
-
-    // Statistics: emit library_video_deleted with the reason bucket and a size
-    // snapshot if known. Best-effort; never blocks the delete.
-    try {
-      // Lazy require to avoid import cycles between storageService and statistics.
-      const { recordEvent } = require("../statistics") as {
-        recordEvent: (input: any) => string | null;
-      };
-      const fileSizeBytes =
-        videoToDelete.fileSize !== undefined && videoToDelete.fileSize !== null
-          ? Number(videoToDelete.fileSize) || 0
-          : 0;
-      recordEvent({
-        eventType: "library_video_deleted",
-        actorRole: reason === "retention" ? "system" : "system",
-        surface: "background",
-        videoId: id,
-        platform:
-          typeof videoToDelete.source === "string"
-            ? videoToDelete.source.toLowerCase()
-            : null,
-        payload: {
-          reason,
-          videoTitle: videoToDelete.title,
-          fileSizeBytes,
-        },
-      });
-    } catch {
-      // statistics is best-effort
-    }
-    return true;
-  } catch (error) {
-    logger.error(
-      "Error deleting video",
-      error instanceof Error ? error : new Error(String(error))
-    );
-    throw new DatabaseError(
-      `Failed to delete video: ${id}`,
-      error instanceof Error ? error : new Error(String(error)),
-      "deleteVideo"
-    );
-  }
-}
-
-function isLocalManagedVideo(
-  video: import("./types").Video
-): boolean {
-  return !video.videoPath || video.videoPath.startsWith("/videos/");
-}
-
-export function isThumbnailReferencedByOtherVideo(
-  video: import("./types").Video,
-  exceptionId: string
-): boolean {
-  if (!video.thumbnailFilename && !video.thumbnailPath) {
-    return false;
-  }
-
-  const allVideos = getVideos();
-
-  return allVideos.some((candidate) => {
-    if (candidate.id === exceptionId) {
-      return false;
-    }
-
-    if (video.thumbnailPath && candidate.thumbnailPath === video.thumbnailPath) {
-      return true;
-    }
-
-    return Boolean(
-      video.thumbnailFilename &&
-        candidate.thumbnailFilename === video.thumbnailFilename &&
-        (!video.thumbnailPath ||
-          !candidate.thumbnailPath ||
-          candidate.thumbnailPath === video.thumbnailPath)
-    );
-  });
-}
-
-function deleteVideoFile(
-  video: import("./types").Video,
-  allCollections: import("./types").Collection[]
-): void {
-  if (!isLocalManagedVideo(video)) {
-    return;
-  }
-
-  // Prefer videoPath so a templated nested path (e.g.
-  // /videos/Channel/Season 2026/file.mp4) resolves to its real on-disk
-  // location. Once template subdirectories are enabled identical basenames
-  // can exist in different folders, and findVideoFile() (basename-only
-  // lookup) could match the wrong file or miss the intended one entirely.
-  const resolved = video.videoPath ? resolveManagedWebPath(video.videoPath) : null;
-  if (resolved) {
-    if (deleteLocalVideoPath(resolved.absolutePath, resolved.rootDir)) {
-      return;
-    }
-
-    if (video.videoFilename) {
-      const fallbackPath = findUnambiguousVideoFileFallback(video.videoFilename);
-      if (fallbackPath) {
-        deleteLocalVideoPath(fallbackPath, VIDEOS_DIR);
-      }
-    }
-    return;
-  }
-
-  // Legacy fallback: only basename available (older rows pre-feature).
-  if (video.videoFilename) {
-    const actualPath = findVideoFile(video.videoFilename, allCollections);
-    if (actualPath) {
-      deleteLocalVideoPath(actualPath, VIDEOS_DIR);
-    }
-  }
-}
-
-function deleteLocalVideoPath(absolutePath: string, rootDir: string): boolean {
-  if (!pathExists(absolutePath)) {
-    return false;
-  }
-
-  removeFileIfExists(absolutePath);
-  removeEmptyDirectoryChain(path.dirname(absolutePath), rootDir);
-  return true;
-}
-
-function findUnambiguousVideoFileFallback(filename: string): string | null {
-  const matches = findVideoFilesByFilename(filename);
-  if (matches.length === 1) {
-    return matches[0];
-  }
-
-  if (matches.length > 1) {
-    logger.warn(
-      `Skipping stale-path video deletion because filename is ambiguous: ${filename}`,
-      { matches }
-    );
-  }
-
-  return null;
-}
-
-function deleteThumbnailFile(
-  video: import("./types").Video,
-  allCollections: import("./types").Collection[]
-): void {
-  if (video.thumbnailFilename) {
-    const canUseLocalFallbacks = isLocalManagedVideo(video);
-    let thumbnailPath: string | null = null;
-
-    // Determine the actual file path based on thumbnailPath
-    if (video.thumbnailPath) {
-      if (video.thumbnailPath.startsWith("/videos/")) {
-        // Thumbnail is stored alongside video file
-        thumbnailPath = buildStoragePath(
-          VIDEOS_DIR,
-          video.thumbnailPath.replace(/^\/videos\//, "")
-        );
-      } else if (video.thumbnailPath.startsWith("/images/")) {
-        // Thumbnail is in images directory (may be in collection subdirectory)
-        thumbnailPath = buildStoragePath(
-          UPLOADS_DIR,
-          video.thumbnailPath.replace(/^\//, "")
-        );
-      }
-    }
-
-    // Fallback: try to find by filename if path-based lookup fails
-    // nosemgrep: javascript.pathtraversal.rule-non-literal-fs-filename
-    if (
-      canUseLocalFallbacks &&
-      (!thumbnailPath || !pathExists(thumbnailPath))
-    ) {
-      // Try alongside video file (when moveThumbnailsToVideoFolder is enabled)
-      if (video.videoFilename) {
-        const videoPath = findVideoFile(video.videoFilename, allCollections);
-        if (videoPath) {
-          const videoDir = path.dirname(videoPath);
-          thumbnailPath = buildStoragePath(videoDir, video.thumbnailFilename);
-          if (!pathExists(thumbnailPath)) {
-            thumbnailPath = null;
-          }
-        }
-      }
-    }
-
-    // Final fallback: try standard image locations
-    if (
-      canUseLocalFallbacks &&
-      (!thumbnailPath || !pathExists(thumbnailPath))
-    ) {
-      thumbnailPath = findImageFile(video.thumbnailFilename, allCollections);
-    }
-
-    // Delete the thumbnail file if it exists
-    if (thumbnailPath && pathExists(thumbnailPath)) {
-      if (isThumbnailReferencedByOtherVideo(video, video.id)) {
-        logger.info(
-          `Skipping thumbnail deletion - another video still references it: ${thumbnailPath}`
-        );
-        return;
-      }
-
-      try {
-        const thumbnailWebPath =
-          resolveManagedThumbnailWebPathFromAbsolutePath(thumbnailPath) ||
-          video.thumbnailPath ||
-          null;
-        removeFileIfExists(thumbnailPath);
-        deleteSmallThumbnailMirrorSync(thumbnailWebPath);
-        pruneManagedArtifactParentDirectories(thumbnailPath);
-        logger.info(`Deleted thumbnail file: ${thumbnailPath}`);
-      } catch (error) {
-        logger.error(
-          `Error deleting thumbnail file ${thumbnailPath}`,
-          error instanceof Error ? error : new Error(String(error))
-        );
-      }
-    }
-  }
-}
-
-function deleteAuthorAvatarIfNeeded(
-  video: import("./types").Video,
-  exceptionId: string
-): void {
-  if (video.authorAvatarFilename && video.author) {
-    // Check if there are other videos from the same author
-    const allVideos = getVideos();
-    const otherVideosFromAuthor = allVideos.filter(
-      (v) => v.id !== exceptionId && v.author === video.author
-    );
-
-    // Only delete avatar if this is the last video from this author
-    if (otherVideosFromAuthor.length === 0) {
-      let avatarPath: string | null = null;
-
-      // Determine the actual file path based on authorAvatarPath
-      if (video.authorAvatarPath) {
-        if (video.authorAvatarPath.startsWith("/avatars/")) {
-          // Avatar is in avatars directory
-          avatarPath = buildStoragePath(
-            UPLOADS_DIR,
-            video.authorAvatarPath.replace(/^\//, "")
-          );
-        } else if (video.authorAvatarPath.startsWith("/images/")) {
-          // Legacy: Avatar might be in images directory (for backward compatibility)
-          avatarPath = buildStoragePath(
-            UPLOADS_DIR,
-            video.authorAvatarPath.replace(/^\//, "")
-          );
-        }
-      }
-
-      // Fallback: try to find by filename in avatars directory
-      // nosemgrep: javascript.pathtraversal.rule-non-literal-fs-filename
-      if (!avatarPath || !pathExists(avatarPath)) {
-        const fallbackPath = buildStoragePath(
-          AVATARS_DIR,
-          video.authorAvatarFilename
-        );
-        if (pathExists(fallbackPath)) {
-          avatarPath = fallbackPath;
-        }
-      }
-
-      // Delete the avatar file if it exists
-      if (avatarPath && pathExists(avatarPath)) {
-        try {
-          removeFileIfExists(avatarPath);
-          removeEmptyDirectoryChain(path.dirname(avatarPath), AVATARS_DIR);
-          logger.info(`Deleted author avatar file: ${avatarPath}`);
-        } catch (error) {
-          logger.error(
-            `Error deleting author avatar file ${avatarPath}`,
-            error instanceof Error ? error : new Error(String(error))
-          );
-        }
-      }
-    } else {
-      logger.info(
-        `Skipping avatar deletion - ${otherVideosFromAuthor.length} other video(s) from author "${video.author}" still exist`
-      );
-    }
-  }
-}
-
-function deleteSubtitleFiles(
-  video: import("./types").Video,
-  allCollections: import("./types").Collection[]
-): void {
-  if (video.subtitles && video.subtitles.length > 0) {
-    const canUseLocalFallbacks = isLocalManagedVideo(video);
-    for (const subtitle of video.subtitles) {
-      let subtitlePath =
-        subtitle.path && typeof subtitle.path === "string"
-          ? resolveManagedWebPath(subtitle.path)?.absolutePath ?? null
-          : null;
-
-      // Fallback: try to find by filename if path-based lookup fails
-      // nosemgrep: javascript.pathtraversal.rule-non-literal-fs-filename
-      if (
-        canUseLocalFallbacks &&
-        (!subtitlePath || !pathExists(subtitlePath))
-      ) {
-        // Try root subtitles directory
-        subtitlePath = buildStoragePath(SUBTITLES_DIR, subtitle.filename);
-        if (!pathExists(subtitlePath)) {
-          // Try alongside video file
-          if (video.videoFilename) {
-            const videoPath = findVideoFile(
-              video.videoFilename,
-              allCollections
-            );
-            if (videoPath) {
-              const videoDir = path.dirname(videoPath);
-              subtitlePath = buildStoragePath(videoDir, subtitle.filename);
-            }
-          }
-        }
-      }
-
-      // Delete the subtitle file if it exists
-      if (subtitlePath && pathExists(subtitlePath)) {
-        try {
-          removeFileIfExists(subtitlePath);
-          pruneManagedArtifactParentDirectories(subtitlePath);
-          logger.info(`Deleted subtitle file: ${subtitlePath}`);
-        } catch (error) {
-          logger.error(
-            `Error deleting subtitle file ${subtitlePath}`,
-            error instanceof Error ? error : new Error(String(error))
-          );
-        }
-      }
-    }
-  }
-}
-
-function pruneManagedArtifactParentDirectories(absolutePath: string): void {
-  if (absolutePath.startsWith(VIDEOS_DIR + path.sep)) {
-    removeEmptyDirectoryChain(path.dirname(absolutePath), VIDEOS_DIR);
-    return;
-  }
-
-  if (absolutePath.startsWith(IMAGES_DIR + path.sep)) {
-    removeEmptyDirectoryChain(path.dirname(absolutePath), IMAGES_DIR);
-    return;
-  }
-
-  if (absolutePath.startsWith(SUBTITLES_DIR + path.sep)) {
-    removeEmptyDirectoryChain(path.dirname(absolutePath), SUBTITLES_DIR);
-    return;
-  }
-}
+// Barrel for the video storage layer, split into focused modules.
+// Re-exports the public surface so existing `./videos` imports keep working.
+
+export type { VideoCallerRole } from "./videoQueries";
+export {
+  getVideos,
+  getVideoBySourceUrl,
+  getVideoById,
+  getVideoPartBySourceUrl,
+} from "./videoQueries";
+
+export type { MediaVisibility } from "./videoVisibility";
+export {
+  classifyMediaVisibility,
+  isCloudFileVisibleToVisitor,
+  isVideoPublic,
+} from "./videoVisibility";
+
+export type { SaveVideoOptions } from "./videoMutations";
+export {
+  saveVideo,
+  saveVideoIfAbsent,
+  saveVideoWithInsertFlag,
+  updateVideo,
+} from "./videoMutations";
+
+export { deleteVideo, isThumbnailReferencedByOtherVideo } from "./videoDeletion";
+
+export { formatLegacyFilenames } from "./legacyFilenames";
